remove logical version
