+++ python_modules/dagster/dagster/_core/definitions/decorators/source_asset_decorator.py
@@ -1,9 +1,6 @@
 from typing import TYPE_CHECKING, Mapping, Optional, Sequence, Union, overload
 
-from typing_extensions import Protocol, TypeAlias
-
-from dagster._core.definitions.events import AssetKey, AssetObservation, CoercibleToAssetKeyPrefix
-from dagster._core.definitions.logical_version import LogicalVersion
+from dagster._core.definitions.events import AssetKey, CoercibleToAssetKeyPrefix
 from dagster._core.definitions.metadata import (
     MetadataEntry,
     MetadataUserInput,
@@ -12,8 +9,6 @@ from dagster._core.definitions.metadata import (
 from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.resource_definition import ResourceDefinition
 from dagster._core.definitions.source_asset import SourceAsset, SourceAssetObserveFunction
-from dagster._core.errors import DagsterInvalidObservationError
-from dagster._core.execution.context.compute import OpExecutionContext
 from dagster._core.storage.io_manager import IOManagerDefinition
 
 from ..decorators.op_decorator import _Op
@@ -106,24 +101,8 @@ class _ObservableSourceAsset:
         self.resource_defs = resource_defs
 
     def __call__(self, observe_fn: SourceAssetObserveFunction) -> SourceAsset:
-        if observe_fn is not None:
-            source_asset_name = self.name or observe_fn.__name__
-            source_asset_key = AssetKey([*self.key_prefix, source_asset_name])
-
-            def fn(context: OpExecutionContext) -> None:
-                raw_observation = observe_fn(context)
-                metadata = _raw_observation_to_metadata(raw_observation)
-                context.log_event(
-                    AssetObservation(
-                        asset_key=source_asset_key,
-                        metadata=metadata,
-                    )
-                )
-
-            op = _Op(
-                name="__".join(source_asset_key.path).replace("-", "_"),
-                description=self.description,
-            )(fn)
+        source_asset_name = self.name or observe_fn.__name__
+        source_asset_key = AssetKey([*self.key_prefix, source_asset_name])
 
         return SourceAsset(
             key=source_asset_key,
@@ -136,15 +115,4 @@ class _ObservableSourceAsset:
             group_name=self.group_name,
             resource_defs=self.resource_defs,
             observe_fn=observe_fn,
-        )
-
-
-def _raw_observation_to_metadata(raw_observation: object) -> MetadataUserInput:
-    if isinstance(raw_observation, dict):
-        return raw_observation
-    elif isinstance(raw_observation, LogicalVersion):
-        return {"logical_version": raw_observation}
-    else:
-        raise DagsterInvalidObservationError(
-            "Source asset observe function must return either a metadata dictionary or a `LogicalVersion` instance."
         )
